fLoating point conversion now working

+++ Coursework 2/src/decodeCommands.h
@@ -6,7 +6,7 @@
 
 extern volatile uint64_t* key;
 extern volatile float tar_velocity;
-extern volatile int32_t tar_rotations;
+extern volatile float tar_rotations;
 extern volatile uint32_t motorPWM;
 extern volatile uint32_t err;
